--- src/main/java/im/toduck/domain/persistence/entity/social/SocialCategoryLink.java
@@ -1,30 +1,0 @@
-package im.toduck.domain.persistence.entity.social;
-
-import im.toduck.global.base.entity.BaseEntity;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "social_category_link")
-@NoArgsConstructor
-public class SocialCategoryLink extends BaseEntity {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "social_id", nullable = false)
-	private Social social;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "social_category_id", nullable = false)
-	private SocialCategory socialCategory;
-}
