+++ src/main/java/im/toduck/domain/person/Alarm.java
@@ -1,4 +1,4 @@
-package im.toduck.domain.persistence.entity.person;
+package im.toduck.domain.person;
 
 public enum Alarm {
 	TEN(10),
