+++ src/main/java/im/toduck/domain/person/Day.java
@@ -1,4 +1,4 @@
-package im.toduck.domain.persistence.entity.person;
+package im.toduck.domain.person;
 
 public enum Day {
 	MON,
