+++ src/main/java/im/toduck/domain/person/RepeatDayLink.java
@@ -1,6 +1,5 @@
-package im.toduck.domain.persistence.entity.social;
+package im.toduck.domain.person;
 
-import im.toduck.domain.persistence.entity.user.User;
 import im.toduck.global.base.entity.BaseEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -13,19 +12,23 @@ import jakarta.persistence.Table;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "likes")
+@Table(name = "repeat_day_link")
 @NoArgsConstructor
-public class Like extends BaseEntity {
+public class RepeatDayLink extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id", nullable = false)
-	private User user;
+	@JoinColumn(name = "routine_id")
+	private Routine routine;
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "social_id", nullable = false)
-	private Social social;
+	@JoinColumn(name = "schedule_id")
+	private Schedule schedule;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "repeat_day_id", nullable = false)
+	private RepeatDay repeatDay;
 }
