+++ src/main/java/im/toduck/domain/social/Comment.java
@@ -1,10 +1,9 @@
-package im.toduck.domain.persistence.entity.person;
+package im.toduck.domain.social;
 
-import im.toduck.domain.persistence.entity.social.Social;
-import im.toduck.domain.persistence.entity.user.User;
+import im.toduck.domain.user.persistence.entity.User;
 import im.toduck.global.base.entity.BaseEntity;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -14,19 +13,21 @@ import jakarta.persistence.Table;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "shared_routine")
+@Table(name = "comment")
 @NoArgsConstructor
-public class SharedRoutine extends BaseEntity {
+public class Comment extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	@ManyToOne(fetch = FetchType.LAZY)
+	@Column(nullable = false, columnDefinition = "TEXT")
+	private String content;
+	@ManyToOne
 	@JoinColumn(name = "user_id", nullable = false)
 	private User user;
 
-	@ManyToOne(fetch = FetchType.LAZY)
+	@ManyToOne
 	@JoinColumn(name = "social_id", nullable = false)
 	private Social social;
 }
