+++ src/main/java/im/toduck/domain/social/Social.java
@@ -1,9 +1,12 @@
-package im.toduck.domain.persistence.entity.social;
+package im.toduck.domain.social;
 
-import im.toduck.domain.persistence.entity.user.User;
+import im.toduck.domain.user.persistence.entity.User;
 import im.toduck.global.base.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -13,21 +16,24 @@ import jakarta.persistence.Table;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "comment")
+@Table(name = "social")
 @NoArgsConstructor
-public class Comment extends BaseEntity {
-
+public class Social extends BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	@Column(nullable = false, columnDefinition = "TEXT")
+	@Column(nullable = false, length = 255)
 	private String content;
-	@ManyToOne
+
+	@Enumerated(EnumType.STRING)
+	@Column(nullable = false)
+	private SocialTag tag;
+
+	@Column(nullable = false)
+	private Boolean isAnonymous;
+
+	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "user_id", nullable = false)
 	private User user;
-
-	@ManyToOne
-	@JoinColumn(name = "social_id", nullable = false)
-	private Social social;
 }
