+++ src/main/java/im/toduck/domain/social/SocialImageFile.java
@@ -1,11 +1,8 @@
-package im.toduck.domain.persistence.entity.social;
+package im.toduck.domain.social;
 
-import im.toduck.domain.persistence.entity.user.User;
 import im.toduck.global.base.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -16,25 +13,17 @@ import jakarta.persistence.Table;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "social")
+@Table(name = "social_image_file")
 @NoArgsConstructor
-public class Social extends BaseEntity {
+public class SocialImageFile extends BaseEntity {
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-
-
-	@Column(nullable = false, length = 255)
-	private String content;
-
-	@Enumerated(EnumType.STRING)
-	@Column(nullable = false)
-	private SocialTag tag;
-
-	@Column(nullable = false)
-	private Boolean isAnonymous;
-
+	@Column(nullable = false, length = 1024)
+	private String url;
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id", nullable = false)
-	private User user;
+	@JoinColumn(name = "social_id", nullable = false)
+	private Social social;
+
 }
