+++ src/main/java/im/toduck/domain/user/persistence/entity/Emotion.java
@@ -1,4 +1,4 @@
-package im.toduck.domain.persistence.entity.user;
+package im.toduck.domain.user.persistence.entity;
 
 //TODO : 임의 데이터 변경 필요
 public enum Emotion {
