+++ src/main/java/im/toduck/domain/user/persistence/entity/Follow.java
@@ -1,7 +1,6 @@
-package im.toduck.domain.persistence.entity.social;
+package im.toduck.domain.user.persistence.entity;
 
 import im.toduck.global.base.entity.BaseEntity;
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -13,17 +12,19 @@ import jakarta.persistence.Table;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "social_image_file")
+@Table(name = "follow")
 @NoArgsConstructor
-public class SocialImageFile extends BaseEntity {
+public class Follow extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-	@Column(nullable = false, length = 1024)
-	private String url;
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "social_id", nullable = false)
-	private Social social;
 
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "followed_id", nullable = false)
+	private User followed;
+
+	@ManyToOne
+	@JoinColumn(name = "follower_id", nullable = false)
+	private User follower;
 }
