+++ src/main/java/im/toduck/domain/user/persistence/entity/Record.java
@@ -1,25 +1,42 @@
-package im.toduck.domain.persistence.entity.person;
+package im.toduck.domain.user.persistence.entity;
+
+import java.time.LocalTime;
 
 import im.toduck.global.base.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "repeat_day")
+@Table(name = "record")
 @NoArgsConstructor
-public class RepeatDay extends BaseEntity {
+public class Record extends BaseEntity {
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
+	@Column(nullable = false)
+	private LocalTime time;
+
+	@Column(nullable = false, length = 2048)
+	private String diary;
+
 	@Enumerated(EnumType.STRING)
 	@Column(nullable = false)
-	private Day day;
+	private Emotion emotion;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id", nullable = false)
+	private User user;
+
 }
